+++ src/malegislature.rs
@@ -1,0 +1,290 @@
+use crate::Cli;
+use indexmap::IndexMap;
+use log::{debug, info};
+use scraper::{ElementRef, Html, Selector};
+use url::Url;
+
+pub fn get_search_page(cli: &Cli) -> (bool, Url, String) {
+    // Get default search page, parse, and create refiner map
+    info!("Creating refiner map");
+    let refiner_map = create_refiner_map();
+
+    // Construct search URL
+    let mut search_url = Url::parse("https://malegislature.gov/Bills/Search").unwrap();
+
+    // https://malegislature.gov/Bills/Search?SearchTerms=&Page=1
+    // https://malegislature.gov/Bills/Search?SearchTerms=mbta&Page=1
+    let mut do_search;
+    let search_term = match cli.search_term.as_deref() {
+        None => {
+            do_search = false;
+            String::from("")
+        }
+        Some(search_term) => {
+            do_search = true;
+            search_term.to_string()
+        }
+    };
+    search_url
+        .query_pairs_mut()
+        .append_pair("SearchTerms", search_term.as_str())
+        .append_pair("Page", "1");
+
+    // https://malegislature.gov/Bills/Search
+    // https://malegislature.gov/Bills/Search?SearchTerms=&Page=1&Refinements%5Blawsgeneralcourt%5D=3139326e64202832303231202d203230323229
+    print_entries_or_append_query_pair(
+        cli.general_court.as_deref(),
+        refiner_map.get("General Court").unwrap(),
+        &mut String::from("Refinements[lawsgeneralcourt]"),
+        &mut search_url,
+    );
+
+    // https://malegislature.gov/Bills/Search?SearchTerms=&Page=1&Refinements%5Blawsbranchname%5D=486f757365
+    do_search = match print_entries_or_append_query_pair(
+        cli.branch.as_deref(),
+        refiner_map.get("Branch").unwrap(),
+        &mut String::from("Refinements[lawsbranchname]"),
+        &mut search_url,
+    ) {
+        None => do_search,
+        Some(do_search) => do_search,
+    };
+
+    // https://malegislature.gov/Bills/Search?SearchTerms=&Page=1&Refinements%5Blawsuserprimarysponsorname%5D=4172636965726f2c204a616d6573
+    do_search = match print_entries_or_append_query_pair(
+        cli.sponsor_legislator.as_deref(),
+        refiner_map.get("Sponsor — Legislator").unwrap(),
+        &mut String::from("Refinements[lawsuserprimarysponsorname]"),
+        &mut search_url,
+    ) {
+        None => do_search,
+        Some(do_search) => do_search,
+    };
+
+    // https://malegislature.gov/Bills/Search?SearchTerms=&Page=1&Refinements%5Blawscommitteeprimarysponsorname%5D=3139326e64204a52756c6573
+    do_search = match print_entries_or_append_query_pair(
+        cli.sponsor_committee.as_deref(),
+        refiner_map.get("Sponsor — Committee").unwrap(),
+        &mut String::from("Refinements[lawscommitteeprimarysponsorname]"),
+        &mut search_url,
+    ) {
+        None => do_search,
+        Some(do_search) => do_search,
+    };
+
+    // https://malegislature.gov/Bills/Search?SearchTerms=&Page=1&Refinements%5Blawsotherprimarysponsorname%5D=41756469746f72206f662074686520436f6d6d6f6e7765616c7468
+    do_search = match print_entries_or_append_query_pair(
+        cli.sponsor_other.as_deref(),
+        refiner_map.get("Sponsor — Other").unwrap(),
+        &mut String::from("Refinements[lawsotherprimarysponsorname]"),
+        &mut search_url,
+    ) {
+        None => do_search,
+        Some(do_search) => do_search,
+    };
+
+    // https://malegislature.gov/Bills/Search?SearchTerms=&Page=1&Refinements%5Blawsfilingtype%5D=416d656e646d6%%6e74
+    do_search = match print_entries_or_append_query_pair(
+        cli.document_type.as_deref(),
+        refiner_map.get("Document Type").unwrap(),
+        &mut String::from("Refinements[lawsfilingtype]"),
+        &mut search_url,
+    ) {
+        None => do_search,
+        Some(do_search) => do_search,
+    };
+    (do_search, search_url, search_term)
+}
+
+#[derive(Debug, Clone)]
+pub struct RefinerEntry {
+    pub refiner_label: String,
+    pub refiner_token: String,
+}
+pub fn create_refiner_map() -> IndexMap<String, IndexMap<String, RefinerEntry>> {
+    // Use an IndexMap to preserve order
+    let mut refiner_map = IndexMap::new();
+
+    // Get the page from which to parse refiners
+    let body = reqwest::blocking::get("https://malegislature.gov/Bills/Search?SearchTerms=&Page=1")
+        .unwrap()
+        .text()
+        .unwrap();
+    let document = Html::parse_document(body.as_str());
+
+    // Define all selectors required to select the refiners
+    let refiner_selector = Selector::parse("div#refiners").unwrap();
+    let group_selector = Selector::parse("fieldset").unwrap();
+    let title_selector = Selector::parse("h4.modal-title").unwrap();
+    let legend_selector = Selector::parse("legend").unwrap();
+    let body_selector = Selector::parse("div.modal-body").unwrap();
+    let label_selector = Selector::parse("label").unwrap();
+    let input_selector = Selector::parse("input").unwrap();
+
+    // Find the div#refiners element which contains all refiner groups, then consider each group
+    let refiner_element = document.select(&refiner_selector).next().unwrap();
+    for group_element in refiner_element.select(&group_selector) {
+        // Use an IndexMap to preserve order
+        let mut refiner_group_map = IndexMap::new();
+
+        // Find the title in the h4.modal-title element ...
+        let group_label_element: ElementRef = match group_element.select(&title_selector).next() {
+            None => {
+                // ... but if it isn't there, find it in the fieldset element
+                group_element.select(&legend_selector).next().unwrap()
+            }
+            Some(element) => element,
+        };
+        let group_label = String::from(group_label_element.text().collect::<Vec<_>>()[0].trim());
+        debug!("\nGroup label: {:?}", group_label);
+
+        // Find the div.modal-body element ...
+        let group_column_element = match group_element.select(&body_selector).next() {
+            None => {
+                // ... but if it isn't there, clone the fieldset element, then consider each row label element
+                group_element.clone()
+            }
+            Some(element) => element,
+        };
+        for row_label_element in group_column_element.select(&label_selector) {
+            // Assign label for this refiner
+            let refiner_label = row_label_element.text().collect::<Vec<_>>()[1]
+                .trim()
+                .replace("  ", " ");
+            debug!("Refiner label: {:?}", refiner_label);
+
+            // Create a unique key from the label
+            let mut refiner_key: String =
+                String::from(refiner_label.split(' ').collect::<Vec<&str>>()[0])
+                    .trim_end_matches(",")
+                    .parse::<String>()
+                    .unwrap()
+                    .replace("'", "-");
+            if !(group_label.contains("Court")
+                || group_label.contains("Branch")
+                || group_label.contains("Legislator"))
+            {
+                let words = refiner_label.split(" ").collect::<Vec<&str>>();
+                refiner_key = words[..words.len() - 1]
+                    .join("-")
+                    .replace("/", "-")
+                    .replace("(", "")
+                    .replace(")", "")
+                    .replace("'", "")
+                    .replace(",", "")
+                    .replace(".", "");
+            }
+            debug!("Refiner key: {:?}", refiner_key);
+
+            // Find the token for this entry
+            let input_element = row_label_element.select(&input_selector).next().unwrap();
+            let refiner_token = input_element.value().attr("data-refinertoken").unwrap();
+            debug!("Refiner token: {}", refiner_token);
+
+            // Collect each refiner group entry key, label, and token
+            refiner_group_map.insert(
+                String::from(refiner_key),
+                RefinerEntry {
+                    refiner_label: String::from(refiner_label),
+                    refiner_token: String::from(refiner_token),
+                },
+            );
+        }
+        refiner_map.insert(String::from(group_label), refiner_group_map);
+    }
+    refiner_map
+}
+
+pub fn print_entries_or_append_query_pair(
+    argument: Option<&str>,
+    refiner_group_map: &IndexMap<String, RefinerEntry>,
+    refiner_field: &mut String,
+    search_url: &mut Url,
+) -> Option<bool> {
+    match argument {
+        Some(refiner_key) if refiner_key == &String::from("MISSING") => {
+            // Refiner key is missing, so list all possible keys
+            for (refiner_key, refiner_entry) in refiner_group_map.iter() {
+                println!(
+                    r#"Use "{}" for "{}""#,
+                    refiner_key, refiner_entry.refiner_label
+                );
+            }
+            None
+        }
+        Some(refiner_key) => {
+            // Refiner key is not missing, so append the query pair
+            search_url.query_pairs_mut().append_pair(
+                refiner_field,
+                refiner_group_map
+                    .get(refiner_key)
+                    .unwrap()
+                    .refiner_token
+                    .as_str(),
+            );
+            Some(true)
+        }
+        None => None,
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct SearchEntry {
+    pub bill_url: Url,
+    pub bill_sponsor: String,
+    pub bill_summary: String,
+}
+pub fn get_and_print_search_results(url: &Url) -> IndexMap<String, SearchEntry> {
+    // Use an IndexMap to preserve order
+    let mut search_results_map = IndexMap::new();
+
+    // Get the search result page, select the table, and parse each result row
+    let body = reqwest::blocking::get(url.clone()).unwrap().text().unwrap();
+    let document = Html::parse_document(body.as_str());
+    let table_body_selector = Selector::parse("tbody").unwrap();
+    let table_row_selector = Selector::parse("tr").unwrap();
+    let table_body_element = document.select(&table_body_selector).next().unwrap();
+    println!("Bill — Link — Sponsor — Summary");
+    // TODO: Handle paging?
+    for table_row_element in table_body_element.select(&table_row_selector) {
+        let (bill_number, bill_url) = get_cell_data(&table_row_element, 2);
+        let (bill_sponsor, _) = get_cell_data(&table_row_element, 3);
+        let (bill_summary, _) = get_cell_data(&table_row_element, 4);
+        println!("{bill_number} — {bill_url} — {bill_sponsor} — {bill_summary}");
+
+        // Collect each search result bill number, url, sponsor, and summary
+        search_results_map.insert(
+            bill_number,
+            SearchEntry {
+                bill_url,
+                bill_sponsor,
+                bill_summary,
+            },
+        );
+    }
+    search_results_map
+}
+
+pub fn get_cell_data(table_row_element: &ElementRef, cell: i32) -> (String, Url) {
+    // Most cell elements contains a hyperlink element ...
+    let base_url = Url::parse("https://malegislature.gov").unwrap();
+    let mut cell_selector = Selector::parse(format!("td:nth-child({cell}) a").as_str()).unwrap();
+    match table_row_element.select(&cell_selector).next() {
+        None => {
+            // ... but if not, use the cell element, otherwise ...
+            cell_selector = Selector::parse(format!("td:nth-child({cell})").as_str()).unwrap();
+            let cell_element = table_row_element.select(&cell_selector).next().unwrap();
+            (
+                String::from(cell_element.text().collect::<Vec<_>>()[0].trim()),
+                base_url,
+            )
+        }
+        Some(cell_element) => (
+            // ... use the hyperlink element
+            String::from(cell_element.text().collect::<Vec<_>>()[0].trim()),
+            base_url
+                .join(cell_element.value().attr("href").unwrap())
+                .unwrap(),
+        ),
+    }
+}
